refactoring landing page code
