+++ components/landing/Partners.tsx
